1.修改步进电机步数长度解析限制； 2.修改电磁阀开关协议值解析； 3.修改直流无刷电机初始化重复流程； 4.添加协议版本信息上报功能。

+++ Core/Inc/app/app_control.h
@@ -128,7 +128,7 @@ void air_supply_valve_off(void);
 //电机控制
 void step_motor_init(void);
 void step_motor_control(StepMotorId_t id, StepMotorCmd_t cmd, uint16_t value);
-void set_step_motor_step_number(StepMotorId_t id, uint16_t steps);
+void set_step_motor_step_number(StepMotorId_t id, uint32_t steps);
 void TimerTask(void *pvParameters); // 定时器任务
 void appInitTask(void *pvParameters); //初始化任务
 void StartTimer(uint8_t func_id, uint32_t timeout_ms);
